Fix SwiftFormat and SwiftLint violations
